Add sixth STRETCH problem to activity
Had a few pairs get through all five problems in the allotted 90
minutes, so I found a suitable sixth problem on CodeWars for them. It
is rather difficult (though one pair *did* manage to solve it), but
should hopefully ensure that no one runs out of things to do in the
future.

Also marked problem 5 as a STRETCH so there is not only one STRETCH
problem, and so that those who had trouble with it don't feel like they
are behind.